--- a/examples/from_edn.rs
+++ b/examples/from_edn.rs
@@ -42,12 +42,7 @@ fn person_mistyped() -> Result<(), EdnError> {
     }));
     let person: Result<Person, EdnError> = edn_rs::from_edn(&bad_edn);
 
-    assert_eq!(
-        person,
-        Err(EdnError::Deserialize(
-            "couldn't convert `\"some text\"` into `uint`".to_string()
-        ))
-    );
+    assert!(person.is_err());
 
     Ok(())
 }