--- a/src/edn/error.rs
+++ b/src/edn/error.rs
@@ -20,6 +20,7 @@ pub enum Code {
 
     /// Parse errors
     HashMapDuplicateKey,
+    SetDuplicateKey,
     InvalidChar,
     InvalidEscape,
     InvalidKeyword,