@Grinkers 2024-02-23
https://github.com/edn-rs/edn-rs/pull/150/commits/3965bc3221d79b883a7059d7ce72c4c5c2eee366
I ended up adding a `CustomError` fallback for crates like `edn-derive`, std-only.

@evaporei 
Matching PR for compatibility over at (this is a breaking change). Please tell me if you have any better ideas. I believe derive will always rely on std, so it's easy enough to add std stuff without messing with the base layer
https://github.com/edn-rs/edn-derive/pull/44