@Grinkers 2024-02-25
Added a new Error for duplicate in sets, matching clojure.