@Grinkers 2024-03-21
> Sorry for the delay, I'm moving to another city and have had limited time. I can take a look as soon as the branch is updated

No worries, I don't think it's too urgent. It would be nice to be merged in so I can tackle the other issues though.  Rebased, which is what I assume you meant. No other changes. Tell me if you want this broken up/squashed in a different way.

I think https://github.com/edn-rs/edn-rs/issues/138 should be a higher priority (github release action -> crates.io publishing). I think we should do a final 0.17 before this PR gets merged too.