@Grinkers 2024-05-27
Code coverage bot seems really bad? It's marking tons of things as untested. 

To compare, here's the source based coverage generated by https://github.com/Grinkers/edn-rs/blob/5ad0427304dc9a05276ef2de36013c82e5ab22e8/bb.edn#L38

[parse.rs.html.zip](https://github.com/edn-rs/edn-rs/files/15454326/parse.rs.html.zip)